Second version update - Updated the database to store the rate per unit as Double instead of Integer in the BillItems table. - The index for the ratePerUnit field changed so modified the Bill Controller - Bill Item class was also modified and the data type of ratePerUnit was changed - Search fields were added to AddProductDialog and ReportMenu - The ratePerUnit field in the billMenu was also updated to handle double values - Selected customer is now being displayed at the billMenu - Total sale for a row is rounded to integer value

+++ src/main/java/Model/BillItem.java
@@ -2,12 +2,12 @@ package Model;
 public class BillItem {
     private final Product product;
     private double quantity;
-    private int ratePerUnit;
+    private double ratePerUnit;
     private String unitType;
     private int total;
     private int discount = 0;
 
-    public BillItem(Product product, double quantity, int ratePerUnit,int discount, String unitType) {
+    public BillItem(Product product, double quantity, double ratePerUnit,int discount, String unitType) {
         this.product = product;
         this.quantity = quantity;
         this.ratePerUnit = ratePerUnit;
@@ -28,7 +28,7 @@ public class BillItem {
         return quantity;
     }
 
-    public int getRatePerUnit() {
+    public double getRatePerUnit() {
         return ratePerUnit;
     }
 
@@ -44,7 +44,7 @@ public class BillItem {
         this.quantity = quantity;
     }
 
-    public void setRatePerUnit(int ratePerUnit) {
+    public void setRatePerUnit(double ratePerUnit) {
         this.ratePerUnit = ratePerUnit;
     }
